extract out CreateLinkConfiguration interactor

Committed-By-Agent: claude
